Add user info component on sidebar
Added a component (userInfo.jsx)that will display the information of the logged-in user on the top of the sidebar. This will include the name of the portal logged in (i.e. Parents, Students, Teachers, Admin, etc), the profile picture of the user logged in, and the name of the user logged in.